fixed a compiler warning
git-svn-id: http://astade.tigris.org/svn/astade/trunk@1703 aaa30fad-34f9-0310-8aa9-877cc762ee2e

+++ Model/package_SVSWQPEV/Package_EWMKSFJL/Package_KYPQHTJS/classes/class_AUHKWVJB/operations/operation_TIGRMATC/code.cpp
@@ -11,7 +11,7 @@ wxCoord x,y;
 do
 {
     dc.GetTextExtent(string.Mid(0,++first),&x,&y);
-} while ((x < pixelCount) && (first <= string.Len()));
+} while ((x < pixelCount) && (first <= static_cast<int>(string.Len())));
 
 wxString firstPart = string.Mid(0,--first);
